Move Pair class to util directory for better organization.
